src/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-course"
version = "0.1.0"
description = "AI Course Django Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alabaster==0.7.16",
    "asgiref==3.8.1",
    "attrs==24.3.0",
    "babel==2.16.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.0",
    "dj-database-url==1.2.0",
    "celery==5.5.1",
    "redis==5.2.1",
    "Django==5.1.4",
    "django-celery-beat==2.8.0",
    "django-redis==5.4.0",
    "django-ratelimit==4.1.0",
    "django-admin-rangefilter==0.10.0",
    "django-admin-extra-buttons==2.1",
    "django-cors-headers==4.6.0",
    "django-filter==22.1",
    "django-jazzmin==3.0.1",
    "django-nonrelated-inlines==0.2",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.2.2",
    "django-tinymce==4.1.0",
    "docutils==0.18.1",
    "drf-spectacular==0.26.1",
    "gunicorn==20.1.0",
    "uvicorn==0.34.2",
    "uvicorn-worker==0.3.0",
    "idna==3.10",
    "imagesize==1.4.1",
    "inflection==0.5.1",
    "Jinja2==3.1.4",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "MarkupSafe==3.0.2",
    "packaging==24.2",
    "psycopg2-binary==2.9.5",
    "Pygments==2.18.0",
    "PyJWT==2.10.1",
    "python-dateutil==2.8.2",
    "python-decouple==3.8",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rpds-py==0.22.3",
    "setuptools==75.6.0",
    "six==1.17.0",
    "snowballstemmer==2.2.0",
    "Sphinx==6.2.1",
    "sphinx-rtd-theme==1.2.1",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jquery==4.1",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sqlparse==0.4.3",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
    "bs4==0.0.2",
    "googletrans==4.0.2",
    "whitenoise==6.8.2",
    "django-admin-sortable==2.3",
    "django-json-widget==2.0.1",
    "markdown==3.7",
    "openai==1.61.1",
    "tiktoken==0.9.0",
    "social-auth-app-django==5.4.3",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "courses", "docs", "locale", "mail", "openai_chats", "payments", "users", "webapp"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
# Django-specific import ordering
known-first-party = ["api", "core", "courses", "docs", "locale", "mail", "openai_chats", "payments", "users", "webapp"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
